pyproject.toml
[project]
name = "flisol"
version = "0.1.0"
description = "FLISoL Blumenau"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml (>=6.0.3,<7.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.4,<0.16.0)"
]

[tool.ruff]
line-length = 120
lint.select = ["I", "E", "F", "W", "C90"]
lint.ignore = ["E501"]

[tool.poetry.scripts]
flisol = "flisol:main"